Move console prompt function to display file
